pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "reference-renamer"
version = "0.1.0"
description = "A tool for standardizing academic article filenames"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Luke Steuber", email = "luke@steuber.io"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing :: General",
]
dependencies = [
    "aiohttp>=3.8.0",
    "backoff>=2.0.0",
    "bibtexparser>=1.4.0",
    "click>=8.0.0",
    "pdf2image>=1.16.0",
    "Pillow>=9.0.0",
    "PyPDF2>=3.0.0",
    "pytesseract>=0.3.10",
    "python-dateutil>=2.8.0",
    "rich>=10.0.0",
    "structlog>=21.0.0",
]

[project.urls]
Homepage = "https://github.com/lukeslp/reference-renamer"
Repository = "https://github.com/lukeslp/reference-renamer"
Issues = "https://github.com/lukeslp/reference-renamer/issues"

[project.scripts]
reference-renamer = "reference_renamer.cli.main:main"
reref = "reference_renamer.cli.main:main"
schollama = "reference_renamer.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["reference_renamer*"]
exclude = ["legacy*", "snippets*"]

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=reference_renamer"
testpaths = [
    "tests",
] 